\name{poly_basis}
\alias{poly_basis}
\title{Polynomial basis}
\usage{poly_basis(x, degree = 3L, intercept = TRUE)}
\arguments{\item{x}{Numeric vector or matrix.}\item{degree}{Maximum marginal degree.}\item{intercept}{Include an intercept.}}
\description{Constructs an additive marginal polynomial basis.}
\value{A numeric matrix.}
