\name{rbf_kernel}
\alias{rbf_kernel}
\title{Gaussian radial-basis kernel}
\usage{rbf_kernel(x, centers, bandwidth)}
\arguments{\item{x}{Numeric vector or matrix.}\item{centers}{Kernel centers.}\item{bandwidth}{Positive bandwidth.}}
\description{Constructs Gaussian RBF features.}
\value{A numeric matrix.}
